requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pillow==10.4.0
numpy==1.26.4
mlflow==2.16.0
boto3==1.34.159
python-multipart
keras==3.10.0
tensorflow